lib.sh: Fix the back-up procedure
Until now the back-up functions gave bad names to the backed up files.
Now it creats a file name which is just the base name of the PREFIX +
the file name itself. It does not guarantee that the names in the
back-up directory will be unique but it should be good enough for now.